+++ admin/customcss.php
@@ -5,7 +5,7 @@
  * Copyright (C) 2005-2012 Regis Houssin        <regis.houssin@inodbox.com>
  * Copyright (C) 2015      Jean-François Ferry	<jfefe@aternatik.fr>
  * Copyright (C) 2022	   Paul Lepont          <paul@kawagency.fr>
- * Copyright (C) 2022-2023 Alexandre Spangaro   <aspangaro@open-dsi.fr>
+ * Copyright (C) 2022-2023 Alexandre Spangaro   <alexandre@inovea-conseil.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -37,7 +37,7 @@ require_once DOL_DOCUMENT_ROOT.'/core/lib/admin.lib.php';
 require_once '../lib/oblyon.lib.php';
 
 // Translations *********************************
-$langs->loadLangs(array('admin', 'oblyon@oblyon', 'opendsi@oblyon'));
+$langs->loadLangs(array('admin', 'oblyon@oblyon', 'inovea@oblyon'));
 
 // Access control *******************************
 if (!$user->admin) {
@@ -73,7 +73,7 @@ print load_fiche_titre($page_name, $linkback);
 
 // Configuration header *************************
 $head						= oblyon_admin_prepare_head();
-print dol_get_fiche_head($head, 'customcss', $langs->trans('Module113900Name'), 0, 'opendsi@oblyon');
+print dol_get_fiche_head($head, 'customcss', $langs->trans('Module113900Name'), 0, 'inovea@oblyon');
 
 $form = new Form($db);
 $formfile = new FormFile($db);