+++ class/actions_oblyon.class.php
@@ -1,6 +1,6 @@
 <?php
 /* Copyright (C) 2022 Paul LEPONT           <paul@kawagency.fr>
-/* Copyright (C) 2022 Alexandre Spangaro    <support@open-dsi.fr>
+/* Copyright (C) 2022 Alexandre Spangaro    <alexandre@inovea-conseil.com>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by